requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
google-adk==0.2.0
google-cloud-firestore==2.19.0
google-cloud-storage==2.18.0
pydantic==2.7.0
pydantic-settings==2.3.0
